LPD-78374 avoid exception

+++ modules/dxp/apps/antivirus/antivirus-async-store/src/main/java/com/liferay/antivirus/async/store/AntivirusScannerHelper.java
@@ -152,9 +152,11 @@ public class AntivirusScannerHelper {
 							_dlAppLocalService.deleteFileEntry(classPK);
 						}
 						else {
-							_dlFileEntryLocalService.deleteFileVersion(
-								dlFileEntry.getUserId(),
-								dlFileEntry.getFileEntryId(), version);
+							if (!version.contains("PWC")) {
+								_dlFileEntryLocalService.deleteFileVersion(
+									dlFileEntry.getUserId(),
+									dlFileEntry.getFileEntryId(), version);
+							}
 						}
 
 						_store.deleteFile(
@@ -194,7 +196,6 @@ public class AntivirusScannerHelper {
 						ServiceContext serviceContext = new ServiceContext();
 
 						serviceContext.setCompanyId(companyId);
-
 						serviceContext.setUuid(dlFileEntry.getUuid());
 
 						_userNotificationEventLocalService.
